Move ShaderProgram class to Velaptor.OpenGL.Shaders namespace & folder
