+++ Testing/VelaptorTests/Fakes/ShaderProgramFake.cs
@@ -5,8 +5,8 @@
 namespace VelaptorTests.Fakes
 {
     using Velaptor.NativeInterop.OpenGL;
-    using Velaptor.OpenGL;
     using Velaptor.OpenGL.Services;
+    using Velaptor.OpenGL.Shaders;
     using Velaptor.Reactables.Core;
     using Velaptor.Reactables.ReactableData;
